Decoupled factory bootstrap information from the lwm2m context
I've moved the factory bootstrap information struct out of the
lwm2m context and made it up to the application to pass a bootstrap
config in by a new function Lwm2mCore_SetFactoryBootstrap()

This will allow non-posix operating systems to also make use of the
factory bootstrap mode.

This change as introduced a little duplication in the Bootstrap server
code. Which will be addressed in following patches.

Signed-off-by: Chris Dewbery <Christopher.Dewbery@imgtec.com>