Add support for the Cobertura code-coverage Jenkins plugin.
This patch adds support for the Cobertura Jenkins plugin, which provides useful visualisation of code coverage, and can enforce thresholds that can be used to control the CI builds. Gcov/lcov output is converted to a compatible format by a third-party python script, used and redistributed under the Apache 2.0 license.

Ref: N/A
Signed-off-by: David Antliff <david.antliff@imgtec.com>