Fix error when attempting to read executable resource /7/0/6
This patch fixes undefined behaviour when an attempt is made to create an instance of object 7, and then a read attempt is made on /7/0/6. In addition, this also adds tests for null pointers in tools_common functions that can lead to a crash with bad tools syntax (e.g. awa-server-write -c IMG -o=/7/0).

Ref: FLOWDM-522
Signed-off-by: David Antliff <david.antliff@imgtec.com>