Add support for out-of-memory tests.
This patch defines wrappers for the system malloc and free functions. These automatically override the runtime library's malloc/free. The global variable mockMallocFailCounter can be set to a value to cause a malloc to fail - the counter is decremented for each malloc encountered, and when it reaches zero it will fail that malloc (it will return NULL).

Signed-off-by: David Antliff <david.antliff@imgtec.com>