Add AWA_OPAQUE macro to Awa static API
The opaque data type is a little different from all the others
being that it is variable length (and doesn't provide another
mechanism to determine it's length, i.e NULL terminated)

This means we need to store a provide a pointer to data, as
well as a pointer to store the length to the
AwaStaticClient_RegisterResourceWithPointer() function.

Now we could simply pass in a pointer to a AwaOpaque data type,
but this is setup for dynamic memory allocation, which goes
against the whole point of the WithPointer() function.

So, to allow variable length opaque data types to be statically
allocated, I have added the AWA_OPAQUE macro.

This macro can be used in place of a variable declaration.

for example:

    AWA_OPAQUE(opaque, 32);

    strcpy(opaque.Data, "hello");
    opaque.Size = strlen("hello");

    AWA_OPAQUE(opaque_array, 32)[10];

    strcpy(opaque_array[2].Data, "hello2");
    opaque_array[2].Size = strlen("hello2");

I have also added the internal data type AwaStaticOpaque which allows
the core to access the elements like so:

    AwaStaticOpaque * temp = (AwaStaticOpaque*)&opaque
    printf("size %d value %s\n", temp->Size, temp->Data);

Signed-off-by: Chris Dewbery <Christopher.Dewbery@imgtec.com>